[RFC PATCH] i2c-imx.c: Add support for I2C bus fault recovery
In the event that a I2C bus is disturbed, for instance by a slave missing
a clock pulse, it is desirable to have a way to get the bus back working
other than by power-cycling the whole system. This patch makes it possible to
to have a special function in board support code issue an I2C reset, since
the IMX peripheral is not capable of doing this, and it needs to be done
by bit-banging the corresponding pins in GPIO mode.
The reset function needs to check if the bus is hung by checking the state
of SDA and issue a reset by pulsing SCL a few times as long as SDA is low.

Signed-off-by: David Jander <david@xxxxxxxxxxx>

Conflicts:
	drivers/i2c/busses/i2c-imx.c