Wandboard : Add SDHCI devices 'backwards'.
This so the boot device (SD3) will get denoted mmc0, regardless if there is SD1/SD2 present.
By default an SD-card in SD1 would be mmc0, and if no card in SD1 then SD3 could be mmc0.

This caused difficulties mounting rootfs, since the card in SD1 might not contain a linux rootfs.