Wandboard : Add support for the Wand baseboard, including sgtl5000 audio.
There are a few functions on the baseboard rather than on the CPU module.
This division between module and baseboard is represented in code by a
baseboard file. If anyone makes their own (possibly custom) baseboard,
they could add another baseboard file instead of having a modified version
of the module board file.

Hopefully these baseboard files can be (one day) loaded as modules, so
switchen between EDM modules and baseboards becomes really easy.

This patch adds the file baseboard-wand with (initially) support for the
sgtl5000 codec.

Loading the baseboard-wand as a module is largerly untested (it is possible,
but whether sgtl5000 audio works or not in that setting, is not known).